--- frontend/src/Login&SignUp/Login.css
@@ -1,132 +1,0 @@
-.login-page {
-  text-align: center;
-}
-
-.header {
-  background-color: #fff;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.header .logo {
-  max-width: 150px;
-  margin-bottom: 10px;
-}
-
-.login-bar {
-  background-color: #336633;
-  width: 100%;
-  padding: 10px 0;
-  color: #fff;
-  font-size: 18px;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.login-container {
-  max-width: 400px;
-  margin: 20px auto;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 10px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  background-color: #fff;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.form-group {
-  margin-bottom: 15px;
-  text-align: left;
-  width: 100%;
-}
-
-label {
-  display: block;
-  margin-bottom: 5px;
-  color: #336633;
-}
-
-input[type="text"],
-input[type="password"] {
-  width: calc(100% - 20px);
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  box-sizing: border-box;
-}
-
-.login-button-group {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-.error {
-  color: red;
-  font-size: 12px;
-  display: block;
-  margin-top: 5px;
-  width: 100%;
-  text-align: left;
-}
-
-.login-button {
-  padding: 10px 20px;
-  background-color: #336633;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  margin-bottom: 5px;
-}
-
-.signup-link {
-  font-size: 14px;
-  color: #336633;
-  cursor: pointer;
-  text-decoration: underline;
-  margin-top: -5px;
-}
-
-/* 팝업 스타일 */
-.popup {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #fff;
-  padding: 20px;
-  border-radius: 10px;
-  text-align: center;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-}
-
-.popup-content button {
-  margin-top: 20px;
-  padding: 10px 20px;
-  background-color: #336633;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-/* Dim background effect */
-.dimmed {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 999;
-}
